# doc/openapi/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: iVend Fiscalization Service API
  description: Api for creating, processing, storing receipts
  license:
    name: MIT
servers:
  - url: http://petstore.swagger.io/v1
paths:
  /receipt:
    post:
      summary: Handle incoming receipt
      operationId: createReceipt
      description: "Creates receipt in the system, asynchronously trying to fiscal it"
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/ReceiptDTO'
      tags:
        - receipts
      responses:
        '200':
          description: Null response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ReceiptDTO'
        '500':
          description: Internal server error
components:
  schemas:
    ReceiptDTO:
      type: object
      required:
        - email
        - sno
        - inn
        - place
        - itemName
        - itemPrice
        - paymentType
        - status
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        email:
          type: "string"
          description: "email"
        sno:
          type: "string"
          description: "Система налогооблажения"
          enum:
          - "osn"
          - "usn_income"
          - "usn_income_outcome"
          - "envd"
          - "esn"
          - "patent"
        inn:
          type: "string"
          description: "ИНН, 12 символов"
        place:
          type: "string"
          description: "Место расчетов"
        itemName:
          type: "string"
          description: "Имя товара"
        itemPrice:
          type: "string"
          description: "Цена товара"
        paymentType:
          $ref: '#/Definitions/PaymentType'
        status:
          type: "string"
          description: "Цена товара"
          enum: &receipt_status
          - "PENDING"
          - "SUCCESS"
          - "ERROR"
        createdAt:
          type: "string"
          description: "Milliseconds since UTC"